requirements.txt
# Core dependencies
requests>=2.28.0
termcolor>=2.3.0
dnspython>=2.4.0
psutil>=5.9.0
pyyaml>=6.0.1
tabulate>=0.9.0
colorama>=0.4.6
tqdm>=4.65.0
python-dateutil>=2.8.2
schedule>=1.2.0
rich>=13.3.0
prettytable>=3.7.0
cryptography>=41.0.0
aiohttp>=3.8.5
asyncio>=3.4.3

# Optional dependencies for enhanced features
# Uncomment if needed for your platform
# pdfkit>=1.0.0  # For PDF export support
# matplotlib>=3.7.0  # For chart generation
# pandas>=2.0.0  # For advanced data analysis

# Platform-specific dependencies
# Windows only
pywin32>=305; sys_platform == 'win32'
winregistry>=0.8.3; sys_platform == 'win32'

# Linux/Unix only
# psutil is already included above and works on all platforms
python-xlib>=0.33; sys_platform == 'linux'
dbus-python>=1.3.2; sys_platform == 'linux'

# Android/Termux specific - these are minimal versions that work well on Android
# Note: Some features might be limited on Termux
# These will be ignored on other platforms
# termux-api>=0.0.1; platform_system == 'Linux' and platform_machine == 'aarch64'
